fix(desktop): let assistant read comments by id
