+++ frontend/apps/cli/src/commands/document.ts
@@ -25,7 +25,7 @@ import {unpackHmId} from '@shm/shared/utils/entity-id-url'
 import {hmIdPathToEntityQueryPath} from '@shm/shared/utils/path-api'
 import {getClient, getServerUrl, getOutputFormat, isPretty} from '../index'
 import {formatOutput, renderMarkdown, printError, printSuccess, printInfo, printWarning} from '../output'
-import {documentToMarkdown} from '../markdown'
+import {documentToMarkdown} from '@shm/shared/hm-markdown'
 import {resolveKey} from '../utils/keyring'
 import {resolveIdWithClient} from '../utils/resolve-id'
 import {createSignerFromKey} from '../utils/signer'
