+++ frontend/apps/cli/src/commands/draft.ts
@@ -27,7 +27,7 @@ import * as readline from 'readline'
 import {readInput, mergeMetadata} from './document'
 import {getOutputFormat, isPretty} from '../index'
 import {formatOutput, renderMarkdown, printError, printSuccess, printInfo, printWarning} from '../output'
-import {documentToMarkdown} from '../markdown'
+import {documentToMarkdown} from '@shm/shared/hm-markdown'
 import {parseMarkdown} from '../utils/markdown'
 import {blocksToMarkdown, slugify, draftFilename, parseDraftFilename} from '@seed-hypermedia/client'
 import {editorBlocksToHMBlockNodes} from '@seed-hypermedia/client/editorblock-to-hmblock'
